[PGEC-84] fix: touchpoints refactoring
